+++ create-react-express/client/src/components/Nav/Nav.js
@@ -1,1 +1,11 @@
-export { default } from "./Nav";
+import React from "react";
+
+const Nav = () => (
+  <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
+    <a className="navbar-brand" href="/">
+      React Reading List
+    </a>
+  </nav>
+);
+
+export default Nav;
