+++ create-react-express/client/src/components/Nav/index.js
@@ -1,11 +1,1 @@
-import React from "react";
-
-const Nav = () => (
-  <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
-    <a className="navbar-brand" href="/">
-      React Reading List
-    </a>
-  </nav>
-);
-
-export default Nav;
+export { default } from "./Nav";
